feat(solutions-store): improves architecture
